requirements.txt
click==4.1
dj-database-url==0.3.0
Django==1.7.9
django-classy-tags==0.6.2
django-cms==3.1.2
django-dotenv==1.3.0
django-reversion==1.8.5
django-sekizai==0.8.2
Django-Select2==4.3.1
django-treebeard==3.0
djangocms-admin-style==0.2.7
djangocms-column==1.5
djangocms-file==0.1
djangocms-flash==0.2.0
djangocms-googlemap==0.3
djangocms-inherit==0.1
djangocms-link==1.6.2
djangocms-picture==0.1
djangocms-style==1.5
djangocms-teaser==0.1
djangocms-text-ckeditor==2.6.0
djangocms-video==0.1
html5lib==0.999999
Pillow==2.9.0
pytz==2015.4
six==1.9.0
tzlocal==1.2
wheel==0.24.0
